@@ -1,6 +1,6 @@
 {
   "name": "vue-flatpickr-component",
-  "version": "8.1.6",
+  "version": "9.0.0",
   "description": "Vue.js component for Flatpickr date-time picker",
   "main": "dist/vue-flatpickr.js",
   "browser": "dist/vue-flatpickr.min.js",
@@ -39,9 +39,9 @@
   },
   "devDependencies": {
     "@babel/core": "^7.11.4",
-    "@babel/plugin-transform-object-assign": "^7.10.4",
     "@babel/preset-env": "^7.11.0",
-    "@vue/test-utils": "1.0.5",
+    "@vue/compiler-sfc": "^3.0.0",
+    "@vue/test-utils": "^2.0.0-beta.5",
     "babel-jest": "^26.3.0",
     "babel-loader": "^8.1.0",
     "bootstrap": "^4.5.2",
@@ -54,19 +54,18 @@
     "mini-css-extract-plugin": "^0.11.0",
     "style-loader": "^1.2.1",
     "unminified-webpack-plugin": "^2.0.0",
-    "vue": "^2.6.12",
-    "vue-loader": "^15.9.3",
-    "vue-template-compiler": "^2.6.12",
+    "vue": "^3.0.0",
+    "vue-jest": "^5.0.0-alpha.4",
+    "vue-loader": "^16.0.0-beta.7",
     "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.0"
   },
   "peerDependencies": {
-    "vue": "^2.0.0"
+    "vue": "^3.0.0"
   },
   "engines": {
-    "node": ">= 4.2.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 10.13.0"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -79,9 +78,6 @@
     "transform": {
       "^.+\\.js$": "babel-jest"
     },
-    "setupFiles": [
-      "<rootDir>/__test__/setup.js"
-    ],
     "collectCoverage": true,
     "testURL": "http://localhost",
     "watchPathIgnorePatterns": [
