@@ -1,13 +1,10 @@
 'use strict';
 
-import Vue from 'vue';
+import {createApp} from 'vue';
 
-// Just for demo
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 import App from './App.vue';
 
-new Vue({
-  el: '#app',
-  render: h => h(App),
-});
+createApp(App).mount('#app')
+
